+++ src/Dynamic.h
@@ -2,5 +2,5 @@
 
 class Dynamic {
   public:
-    static uint32_t Knapsack(uint8_t n, uint16_t M, Item *items, TaskType type, int32_t B);
+    static uint32_t Knapsack(uint8_t n, int32_t M, Item *items, TaskType type, int32_t B);
 };
